fix: 피드 InfiniteFeedList 컴포넌트 FeedList 컴포넌트로 통합 수정 #147
